organização de componentes + estilização
