+++ planilhas expositores/Amostras novas - 20_10.csv
@@ -1,0 +1,26 @@
+Laranjal,,,,,
+Codigo ,Expositor,Posicao,Unidade,Referência,Quantidade
+1129982,x,y,Metro Quadrado,,1
+1129699,x,y,Metro Quadrado,,1
+1129115,x,y,Metro Quadrado,,1
+1130848,x,y,Metro Quadrado,,1
+1131000,x,y,Metro Quadrado,,1
+1103485,x,y,Metro Quadrado,,1
+1130765,x,y,Metro Quadrado,,1
+1088208,x,y,Metro Quadrado,,1
+1125957,x,y,Metro Quadrado,,1
+1054857,x,y,Metro Quadrado,,1
+1062455,x,y,Metro Quadrado,,1
+1130495,x,y,Metro Quadrado,,1
+1083305,x,y,Metro Quadrado,,1
+1099373,x,y,Metro Quadrado,,1
+1079748,x,y,Metro Quadrado,,1
+1126461,x,y,Metro Quadrado,,1
+1124872,x,y,Metro Quadrado,,1
+1106194,x,y,Metro Quadrado,,1
+1122010,x,y,Metro Quadrado,,1
+1125834,x,y,Metro Quadrado,,1
+1124806,x,y,Metro Quadrado,,1
+1118346,x,y,Metro Quadrado,,1
+1059435,x,y,Metro Quadrado,,1
+1117652,x,y,Metro Quadrado,,1
