--- planilhas expositores/NOVOS 20_10.csv
@@ -1,26 +1,0 @@
-Laranjal,,,,,
-Codigo,Expositor,Posicao,Unidade,Referencia,Quantidade 
-1129982,,,,,
-1129699,,,,,
-1129115,,,,,
-1130848,,,,,
-1131000,,,,,
-1103485,,,,,
-1130765,,,,,
-1088208,,,,,
-1125957,,,,,
-1054857,,,,,
-1062455,,,,,
-1130495,,,,,
-1083305,,,,,
-1099373,,,,,
-1079748,,,,,
-1126461,,,,,
-1124872,,,,,
-1106194,,,,,
-1122010,,,,,
-1125834,,,,,
-1124806,,,,,
-1118346,,,,,
-1059435,,,,,
-1117652,,,,,
