+++ src/App.css
@@ -33,5 +33,5 @@ button {
 
     .etiquetasContainer button {
         display: none;
-    }    
+    }
 }
