+++ src/components/EtiquetasPromo.module.css
@@ -191,34 +191,4 @@
         right: 2mm;
         font-size: x-small;
     }
-   
-/*    
-    
-    .etiquetasPromoWrapper button {
-        display: block;
-    
-        margin: 0 auto;
-        padding: 0.5rem 1rem;
-        border: none;
-        border-radius: 0.5rem;
-        background-color: #f0f0f0;
-        color: #000;
-        font-size: 1rem;
-        font-weight: 600;
-        cursor: pointer;
-    
-        transition: all 0.2s ease-in-out;
-        
-        &:hover {
-            background-color: #e0e0e0;
-        }
-    
-        &:active {
-            background-color: #d0d0d0;
-        }
-    
-        &:focus {
-            outline: none;
-        }
-    } */
 }
